[FU-237] New Relic Integration (#56)
* FU-237 chore: New Relic configure gradle

* FU-237 chore: update start.sh

배포 환경에서 start.sh으로 애플리케이션 실행할 때 뉴렐릭 에이전트도 함께 실행되도록 스크립트 수정했음

* FU-237 chore: cd_workflow 전략 임시 수정

배포 환경에서 뉴렐릭 에이전트가 잘 실행되는지 확인하고자 함. 병합하기 전에 원복예정

* FU-237 chore: trigger github action

환경변수 세팅이 안되어있어서 등록 후 다시 github action 동작

* FU-237 chore: 환경변수 설정문제 테스트

* FU-237 chore: 환경변수 추가 설정

* FU-237 chore: 환경변수 참조 수정

* FU-237 chore: start.sh 주석 삭제

* FU-237 chore: cd workflow 원복

뉴렐릭 통합 테스트가 끝나서 원복함

+++ scripts/start.sh
@@ -8,13 +8,15 @@ set +o allexport
 # 경로 및 파일 설정
 ROOT_PATH="/home/ubuntu/freebe-backend"
 JAR="$ROOT_PATH/application.jar"
+NEW_RELIC_CONFIG_FILE="$ROOT_PATH/newrelic/newrelic.yml"
+NEW_RELIC_JAR_FILE="$ROOT_PATH/newrelic/newrelic.jar"
 
 APP_LOG="$ROOT_PATH/application.log"
 ERROR_LOG="$ROOT_PATH/error.log"
 START_LOG="$ROOT_PATH/start.log"
 NOW=$(date "+%Y %b %d %a %H:%M:%S")
 
-echo "--------------------------------------------------" >> $START_LOG
+echo "---------------------------------------------------" >> $START_LOG
 
 # JAR 파일 복사
 echo "[$NOW] $JAR 복사 중..." >> $START_LOG
@@ -25,11 +27,17 @@ else
   exit 1
 fi
 
+# New Relic YML 파일이 존재하는지 확인
+if [ ! -f "$NEW_RELIC_CONFIG_FILE" ]; then
+    echo "New Relic 설정 파일이 존재하지 않습니다." >> $START_LOG
+    exit 1
+fi
+
 # 애플리케이션 실행
 echo "[$NOW] > $JAR 실행" >> $START_LOG
-nohup java -jar $JAR > $APP_LOG 2> $ERROR_LOG &
+nohup java -javaagent:$NEW_RELIC_JAR_FILE -Dnewrelic.config.license_key=$NEW_RELIC_LICENSE_KEY -Dnewrelic.config.app_name=$NEW_RELIC_APP_NAME -jar $JAR > $APP_LOG 2> $ERROR_LOG &
 
-# 실행된 프로세스의 PID 확인
+  # 실행된 프로세스의 PID 확인
 NEW_SERVICE_PID=$(pgrep -f $JAR)
 if [ ! -z "$NEW_SERVICE_PID" ]; then
   echo "[$NOW] > 새로운 서비스 PID: $NEW_SERVICE_PID" >> $START_LOG
